Creating a new repository for mandelbrot set visualisation as the last project got a bit out of hand and became sprawling and cumbersome. The goal of this repository is to provide a mandelbrot image generator written in C.
There may be some outstanding design work to think about but the core purpose is to generate a buffer of rgb values corresponding to a visualisation of the mandelbrot set. There are a number of considerations here. How does is colour it? Where does it get the imput to colour it from. What actions will be supported after the buffer is generated? Should real time buffer updates be supported?

One strong idea at present is the option to write to a shared memory block. This shared memory block can be used by another process. Some kind of interaction between the parent process and this program may be necessary.

This repository is not a network server.