Made a large number of changes to server.c.
Created a pipe that can be used in conjunction with epoll so that the
program does not consume excessive cpu while it waits for the routines
generating the image to finish. The main thread of the program watches
the pipe for a message that the image generation is finished then
collects the image generation thread.

Requests for an image to be generated to not block the server anymore.
The callback imediately dispatches a worker thread which the main thread
collects as mentioned above.

While the server is working on building an image, further requests to
build an image will not be processed. Instead the client will recieve a
message that the server is busy.