A number of changes made.
Changing  to uint8 instead of uint16 when transmitting image data to the
image request client node. This is beacause the golang that node is
running is expecting data in uint8 form and uint16 is completely
unecessary.

Added some utitily functions to save the image_data array data as a
file. This could later on be used to quickly provide some legitimate
image data for testing purposes. (ie when testing node communication
having the array data saved to a file saves the time of recalculating
new array data.

Added some logging statements.